+++ packages/create-sitecore-jss/src/templates/nextjs-xmcloud/src/lib/page-props-factory/plugins/component-themes.ts
@@ -7,7 +7,7 @@ class ComponentThemesPlugin implements Plugin {
   order = 10;
 
   async exec(props: SitecorePageProps) {
-    // Collect FEAAS, BYOC, SXA component themes
+    // Collect SXA component themes
     props.headLinks.push(
       ...getDesignLibraryStylesheetLinks(
         props.layoutData,